+++ TradersExtended.cs
@@ -20,7 +20,7 @@ namespace TradersExtended
     {
         private const string pluginID = "shudnal.TradersExtended";
         private const string pluginName = "Traders Extended";
-        private const string pluginVersion = "1.1.3";
+        private const string pluginVersion = "1.1.4";
 
         private Harmony harmony;
 
@@ -57,6 +57,7 @@ namespace TradersExtended
 
         public static ConfigEntry<string> tradersCustomPrefabs;
         public static ConfigEntry<bool> disableVanillaItems;
+        public static ConfigEntry<float> qualityMultiplier;
 
         public static readonly Dictionary<string, List<TradeableItem>> tradeableItems = new Dictionary<string, List<TradeableItem>>();
         public static readonly Dictionary<string, List<TradeableItem>> sellableItems = new Dictionary<string, List<TradeableItem>>();
@@ -153,6 +154,7 @@ namespace TradersExtended
 
             tradersCustomPrefabs = config("Misc", "Custom traders prefab names", defaultValue: "", "List of custom prefab names of Trader added by mods to control coins. Prefab name, case sensitive, comma separated");
             disableVanillaItems = config("Misc", "Disable vanilla items", defaultValue: false, "Disable vanilla items on traders. Custom traders could or could not work depending on their implementation.");
+            qualityMultiplier = config("Misc", "Quality multiplier", defaultValue: 0.0f, "Quality multiplier for price. Each level of additional quality level adds that percent of price.");
 
             InitCommands();
         }